Module google_recommendationengine1_beta1::api

Structs

GoogleApiHttpBody
Message that represents an arbitrary HTTP body. It should only be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page. This message can be used both in streaming and non-streaming API methods in the request as well as the response. It can be used as a top-level request field, which is convenient if one wants to extract parameters from either the URL or HTTP template into the request fields and also want access to the raw HTTP body. Example: message GetResourceRequest { // A unique request id. string request_id = 1; // The raw HTTP body is bound to this field. google.api.HttpBody http_body = 2; } service ResourceService { rpc GetResource(GetResourceRequest) returns (google.api.HttpBody); rpc UpdateResource(google.api.HttpBody) returns (google.protobuf.Empty); } Example with streaming methods: service CaldavService { rpc GetCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); rpc UpdateCalendar(stream google.api.HttpBody) returns (stream google.api.HttpBody); } Use of this type only changes how the request and response bodies are handled, all other features will continue to work unchanged.

GoogleCloudRecommendationengineV1beta1BigQuerySource
BigQuery source import data from.

GoogleCloudRecommendationengineV1beta1Catalog
The catalog configuration. Next ID: 5.

GoogleCloudRecommendationengineV1beta1CatalogInlineSource
The inline source for the input config for ImportCatalogItems method.

GoogleCloudRecommendationengineV1beta1CatalogItem
CatalogItem captures all metadata information of items to be recommended.

GoogleCloudRecommendationengineV1beta1CatalogItemCategoryHierarchy
Category represents catalog item category hierarchy.

GoogleCloudRecommendationengineV1beta1CatalogItemLevelConfig
Configures the catalog level that users send events to, and the level at which predictions are made.

GoogleCloudRecommendationengineV1beta1CreatePredictionApiKeyRegistrationRequest
Request message for the CreatePredictionApiKeyRegistration method.

GoogleCloudRecommendationengineV1beta1EventDetail
User event details shared by all recommendation types.

GoogleCloudRecommendationengineV1beta1FeatureMap
FeatureMap represents extra features that customers want to include in the recommendation model for catalogs/user events as categorical/numerical features.

GoogleCloudRecommendationengineV1beta1FeatureMapFloatList
A list of float features.

GoogleCloudRecommendationengineV1beta1FeatureMapStringList
A list of string features.

GoogleCloudRecommendationengineV1beta1GcsSource
Google Cloud Storage location for input content. format.

GoogleCloudRecommendationengineV1beta1Image
Catalog item thumbnail/detail image.

GoogleCloudRecommendationengineV1beta1ImportCatalogItemsRequest
Request message for Import methods.

GoogleCloudRecommendationengineV1beta1ImportErrorsConfig
Configuration of destination for Import related errors.

GoogleCloudRecommendationengineV1beta1ImportUserEventsRequest
Request message for the ImportUserEvents request.

GoogleCloudRecommendationengineV1beta1InputConfig
The input config source.

GoogleCloudRecommendationengineV1beta1ListCatalogItemsResponse
Response message for ListCatalogItems method.

GoogleCloudRecommendationengineV1beta1ListCatalogsResponse
Response for ListCatalogs method.

GoogleCloudRecommendationengineV1beta1ListPredictionApiKeyRegistrationsResponse
Response message for the ListPredictionApiKeyRegistrations.

GoogleCloudRecommendationengineV1beta1ListUserEventsResponse
Response message for ListUserEvents method.

GoogleCloudRecommendationengineV1beta1PredictRequest
Request message for Predict method.

GoogleCloudRecommendationengineV1beta1PredictResponse
Response message for predict method.

GoogleCloudRecommendationengineV1beta1PredictResponsePredictionResult
PredictionResult represents the recommendation prediction results.

GoogleCloudRecommendationengineV1beta1PredictionApiKeyRegistration
Registered Api Key.

GoogleCloudRecommendationengineV1beta1ProductCatalogItem
ProductCatalogItem captures item metadata specific to retail products.

GoogleCloudRecommendationengineV1beta1ProductCatalogItemExactPrice
Exact product price.

GoogleCloudRecommendationengineV1beta1ProductCatalogItemPriceRange
Product price range when there are a range of prices for different variations of the same product.

GoogleCloudRecommendationengineV1beta1ProductDetail
Detailed product information associated with a user event.

GoogleCloudRecommendationengineV1beta1ProductEventDetail
ProductEventDetail captures user event information specific to retail products.

GoogleCloudRecommendationengineV1beta1PurchaseTransaction
A transaction represents the entire purchase transaction.

GoogleCloudRecommendationengineV1beta1PurgeUserEventsRequest
Request message for PurgeUserEvents method.

GoogleCloudRecommendationengineV1beta1RejoinUserEventsRequest
Request message for CatalogRejoin method.

GoogleCloudRecommendationengineV1beta1UserEvent
UserEvent captures all metadata information recommendation engine needs to know about how end users interact with customers' website.

GoogleCloudRecommendationengineV1beta1UserEventInlineSource
The inline source for the input config for ImportUserEvents method.

GoogleCloudRecommendationengineV1beta1UserInfo
Information of end users.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

ProjectLocationCatalogCatalogItemCreateCall
Creates a catalog item.

ProjectLocationCatalogCatalogItemDeleteCall
Deletes a catalog item.

ProjectLocationCatalogCatalogItemGetCall
Gets a specific catalog item.

ProjectLocationCatalogCatalogItemImportCall
Bulk import of multiple catalog items. Request processing may be synchronous. No partial updating supported. Non-existing items will be created. Operation.response is of type ImportResponse. Note that it is possible for a subset of the items to be successfully updated.

ProjectLocationCatalogCatalogItemListCall
Gets a list of catalog items.

ProjectLocationCatalogCatalogItemPatchCall
Updates a catalog item. Partial updating is supported. Non-existing items will be created.

ProjectLocationCatalogEventStoreOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationCatalogEventStoreOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationCatalogEventStorePlacementPredictCall
Makes a recommendation prediction. If using API Key based authentication, the API Key must be registered using the PredictionApiKeyRegistry service. Learn more.

ProjectLocationCatalogEventStorePredictionApiKeyRegistrationCreateCall
Register an API key for use with predict method.

ProjectLocationCatalogEventStorePredictionApiKeyRegistrationDeleteCall
Unregister an apiKey from using for predict method.

ProjectLocationCatalogEventStorePredictionApiKeyRegistrationListCall
List the registered apiKeys for use with predict method.

ProjectLocationCatalogEventStoreUserEventCollectCall
Writes a single user event from the browser. This uses a GET request to due to browser restriction of POST-ing to a 3rd party domain. This method is used only by the Recommendations AI JavaScript pixel. Users should not call this method directly.

ProjectLocationCatalogEventStoreUserEventImportCall
Bulk import of User events. Request processing might be synchronous. Events that already exist are skipped. Use this method for backfilling historical user events. Operation.response is of type ImportResponse. Note that it is possible for a subset of the items to be successfully inserted. Operation.metadata is of type ImportMetadata.

ProjectLocationCatalogEventStoreUserEventListCall
Gets a list of user events within a time range, with potential filtering. The method does not list unjoined user events. Unjoined user event definition: when a user event is ingested from Recommendations AI User Event APIs, the catalog item included in the user event is connected with the current catalog. If a catalog item of the ingested event is not in the current catalog, it could lead to degraded model quality. This is called an unjoined event.

ProjectLocationCatalogEventStoreUserEventPurgeCall
Deletes permanently all user events specified by the filter provided. Depending on the number of events specified by the filter, this operation could take hours or days to complete. To test a filter, use the list command first.

ProjectLocationCatalogEventStoreUserEventRejoinCall
Triggers a user event rejoin operation with latest catalog data. Events will not be annotated with detailed catalog information if catalog item is missing at the time the user event is ingested, and these events are stored as unjoined events with a limited usage on training and serving. This API can be used to trigger a 'join' operation on specified events with latest version of catalog items. It can also be used to correct events joined with wrong catalog items.

ProjectLocationCatalogEventStoreUserEventWriteCall
Writes a single user event.

ProjectLocationCatalogListCall
Lists all the catalog configurations associated with the project.

ProjectLocationCatalogOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationCatalogOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectLocationCatalogPatchCall
Updates the catalog configuration.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the RecommendationsAI hub.

RecommendationsAI
Central instance to access all RecommendationsAI related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.